# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AKIM AI — Аким на 5 часов",
    "description": "Учебная симуляция на синтетических данных. Детерминированные расчёты; AI объясняет готовые факты.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/ai/analyze": {
      "post": {
        "tags": [
          "AI integration"
        ],
        "summary": "Analyze",
        "operationId": "analyze_api_ai_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AnalyzeRequest"
                  },
                  {
                    "$ref": "#/components/schemas/ScenarioAnalyzeRequest"
                  }
                ],
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Analysis"
                    },
                    {
                      "$ref": "#/components/schemas/StructuredAnalysis"
                    }
                  ],
                  "title": "Response Analyze Api Ai Analyze Post"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/explain": {
      "post": {
        "tags": [
          "AI integration"
        ],
        "summary": "Explain",
        "operationId": "explain_api_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/compare": {
      "post": {
        "tags": [
          "AI integration"
        ],
        "summary": "Compare",
        "operationId": "compare_api_ai_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/report/executive-brief": {
      "post": {
        "tags": [
          "AI integration"
        ],
        "summary": "Report",
        "operationId": "report_api_report_executive_brief_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/advice": {
      "post": {
        "tags": [
          "AI integration"
        ],
        "summary": "Advice",
        "operationId": "advice_api_ai_advice_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdviceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdviceResponse": {
        "properties": {
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/AdviceSuggestionResponse"
            },
            "type": "array",
            "maxItems": 3,
            "title": "Suggestions"
          },
          "validation_mode": {
            "type": "string",
            "const": "individual_additions",
            "title": "Validation Mode",
            "default": "individual_additions"
          },
          "base_scenario_id": {
            "type": "string",
            "title": "Base Scenario Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "priority",
          "reason",
          "suggestions",
          "base_scenario_id"
        ],
        "title": "AdviceResponse",
        "description": "Keep the AI participant's snake_case response, with explicit validation context."
      },
      "AdviceSuggestionResponse": {
        "properties": {
          "measure_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Measure Id"
          },
          "district_id": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "esil",
                  "almaty",
                  "saryarka",
                  "baikonur",
                  "nura"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "District Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "measure_id",
          "district_id"
        ],
        "title": "AdviceSuggestionResponse"
      },
      "Analysis": {
        "properties": {
          "source": {
            "type": "string",
            "const": "ai",
            "title": "Source",
            "default": "ai"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths"
          },
          "risks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risks"
          },
          "tradeoffs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tradeoffs"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary",
          "strengths",
          "risks",
          "tradeoffs",
          "recommendations"
        ],
        "title": "Analysis"
      },
      "AnalyzeRequest": {
        "properties": {
          "simulation": {
            "$ref": "#/components/schemas/SimulationInput"
          },
          "question": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "simulation"
        ],
        "title": "AnalyzeRequest"
      },
...
      "CompareRequest": {
        "properties": {
          "originalDecisions": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Originaldecisions"
          },
          "alternativeDecisions": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Alternativedecisions"
          },
          "question": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "originalDecisions",
          "alternativeDecisions"
        ],
        "title": "CompareRequest"
      },
      "CompareResponse": {
        "properties": {
          "originalScenarioId": {
            "type": "string",
            "title": "Originalscenarioid"
          },
          "alternativeScenarioId": {
            "type": "string",
            "title": "Alternativescenarioid"
          },
          "modelVersion": {
            "type": "string",
            "title": "Modelversion"
          },
          "dataChecksum": {
            "type": "string",
            "title": "Datachecksum"
          },
          "analysis": {
            "$ref": "#/components/schemas/StructuredAnalysis"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "originalScenarioId",
          "alternativeScenarioId",
          "modelVersion",
          "dataChecksum",
          "analysis"
        ],
        "title": "CompareResponse"
      },
...
      "Decision": {
        "properties": {
          "measureId": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Measureid"
          },
          "districtId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Districtid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "measureId"
        ],
        "title": "Decision"
      },
...
      "ErrorResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/Issue"
            },
            "type": "array",
            "title": "Errors"
          },
          "score": {
            "type": "null",
            "title": "Score"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message",
          "code"
        ],
        "title": "ErrorResponse"
      },
      "ExplainRequest": {
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Decisions"
          },
          "question": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decisions"
        ],
        "title": "ExplainRequest"
      },
      "FindingResponse": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "evidence"
        ],
        "title": "FindingResponse"
      },
...
      "Issue": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "measureIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Measureids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Issue"
      },
...
      "ScenarioAnalyzeRequest": {
        "properties": {
          "scenario": {
            "$ref": "#/components/schemas/ScenarioFactsInput"
          },
          "question": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scenario"
        ],
        "title": "ScenarioAnalyzeRequest"
      },
      "ScenarioFactsInput": {
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/SummaryDecisionInput"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Decisions"
          }
        },
        "type": "object",
        "required": [
          "decisions"
        ],
        "title": "ScenarioFactsInput"
      },
...
      "SimulationInput": {
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Decisions"
          }
        },
        "type": "object",
        "required": [
          "decisions"
        ],
        "title": "SimulationInput"
      },
...
      "StructuredAnalysis": {
        "properties": {
          "strengths": {
            "items": {
              "$ref": "#/components/schemas/FindingResponse"
            },
            "type": "array",
            "title": "Strengths"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/FindingResponse"
            },
            "type": "array",
            "title": "Risks"
          },
          "tradeoffs": {
            "items": {
              "$ref": "#/components/schemas/FindingResponse"
            },
            "type": "array",
            "title": "Tradeoffs"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "strengths",
          "risks",
          "tradeoffs",
          "recommendations",
          "answer"
        ],
        "title": "StructuredAnalysis"
      },
      "SummaryDecisionInput": {
        "properties": {
          "measureId": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Measureid"
          },
          "districtId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Districtid"
          }
        },
        "type": "object",
        "required": [
          "measureId"
        ],
        "title": "SummaryDecisionInput"
      },
...
    }
  }
}